+++ moviepy/video/io/ffmpeg_reader.py
@@ -422,7 +422,7 @@ class FFmpegInfosParser:
 
                 # get input number, stream number, language and type
                 main_info_match = re.search(
-                    r"^Stream\s#(\d+):(\d+)\(?(\w+)?\)?:\s(\w+):", line.lstrip()
+                    r"^Stream\s#(\d+):(\d+)[\[(]?(\w+)?[\])]?:\s(\w+):", line.lstrip()
                 )
                 (
                     input_number,
@@ -434,12 +434,16 @@ class FFmpegInfosParser:
                 stream_number = int(stream_number)
                 stream_type_lower = stream_type.lower()
 
+                if language is not None:
+                    if language == "und" or language.startswith("0x"):
+                        language = None
+
                 # start builiding the current stream
                 self._current_stream = {
                     "input_number": input_number,
                     "stream_number": stream_number,
                     "stream_type": stream_type_lower,
-                    "language": language if language != "und" else None,
+                    "language": language,
                     "default": not self._default_stream_found
                     or line.endswith("(default)"),
                 }